+++ .eslintrc
@@ -1,8 +1,8 @@
 {
-   "extends": ["react-app", "airbnb", "prettier", "prettier/react"],
-   "plugins": ["prettier"],
-   "rules": {
-      "react/jsx-props-no-spreading": "off",
-      "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
-   }
+  "extends": ["react-app"],
+  "plugins": ["prettier"],
+  "rules": {
+    "react/jsx-props-no-spreading": "off",
+    "react/jsx-filename-extension": [1, { "extensions": [".js", ".jsx"] }]
+  }
 }
