+++ src/App.js
@@ -4,15 +4,14 @@ import Home from './components/Home';
 import About from './components/About';
 import NavBar from './components/NavBar';
 import Footer from './components/Footer';
-import OrganizationContainer from './components/OrganizationContainer';
 import Login from './components/Login';
 import UserContainer from './components/UserContainer';
-import EventContainer from './components/EventContainer';
 import SignUp from './components/SignUp';
-import NewEvent from './components/NewEvent';
-import Event from './components/Event';
-import OrganizationCard from './components/OrganizationCard';
-import EventDetail from './components/EventDetail';
+import NewEvent from './components/Events/NewEvent';
+import EventDetail from './components/Events/EventDetail';
+import Event from './components/Events/Event';
+import OrganizationContainer from './components/Organizations/OrganizationContainer';
+import OrganizationCard from './components/Organizations/OrganizationCard';
 import EditVolunteer from './components/EditVolunteer';
 import './App.css';
